refactor: prevent multiple class declararations
