--- addons/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\zt
