--- addons/config.cpp
@@ -1,8 +1,0 @@
-#include "sequencer\sequencer.cpp"
-
-class CfgPatches {
-    class ZT {
-        requiredVersion = 1.0;
-        requiredAddons[] = {"A3_Modules_F"};
-    };
-};
