+++ addons/sequencer/modules/HelloWorld/CfgFunctions.hpp
@@ -1,0 +1,4 @@
+class HelloWorld {
+	file = "z\ZT\Sequencer\modules\HelloWorld\functions";
+	class helloWorld {};
+};
